Rewards feature
One of the 3 features highlighted in my design: Donation Rewards

I realized that loyalty rewards keep me consistent. I personally keep going to the same grocery store because of the Loyalty Points / Rewards system and I thought that would be a nice way to reward people and motivate them to donate their food when they can as opposed to throwing it out.